% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{eratos_request}
\alias{eratos_request}
\title{Generic API request for Eratos and Senaps}
\usage{
eratos_request(
  type = c("eratos.workspace", "senaps.sensor", "senaps.datasource", "senaps.tmd"),
  path = "/",
  method = "GET",
  query = NULL,
  body = NULL,
  ...
)
}
\arguments{
\item{type}{One of: "eratos.workspace", "senaps.sensor", "senaps.datasource", "senaps.tmd"}

\item{path}{API endpoint path (e.g. "/workspaces" or "streams/list")}

\item{method}{HTTP method, e.g. "GET", "POST", "PUT", "DELETE"}

\item{query}{List of query parameters}

\item{body}{Optional JSON body for POST/PUT requests}
}
\value{
Parsed JSON content or response object
}
\description{
Generic API request for Eratos and Senaps
}
